+++ lib/screens/screens.dart
@@ -1,0 +1,4 @@
+// Export all screens for easy importing
+export 'dashboard/dashboard_screen.dart';
+export 'add_gear/add_gear_screen.dart';
+// TODO: Add more screens as they are implemented